+++ guest.h
@@ -11,9 +11,9 @@
 class Guest:public Peer{
 protected:
     char sourceip[INET6_ADDRSTRLEN];
-    int  sourceport;
+    uint16_t  sourceport;
     char destip[INET6_ADDRSTRLEN];
-    int  destport;
+    uint16_t  destport;
     char rbuff[4096];
     uint32_t  read_len=0;
     uint32_t expectlen=0;
